Move scheduler state into control-plane scheduler
